Move ThresholdMessage to web/pages/reports
